backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-retry==2.9.0
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.2.0
certifi==2023.5.7
cffi==1.17.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==43.0.3
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.19.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
frozenlist==1.3.3
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
identify==2.5.24
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
openai==0.27.8
orjson==3.9.1
passlib==1.7.4
platformdirs==3.7.0
pre-commit==3.3.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.9
PyJWT==2.9.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.8.8
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tiktoken==0.5.1
tqdm==4.65.0
twilio==9.3.6
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
